requirements.txt
numpy==1.22.2
pandas==1.4.2
Pillow==9.0.1
tqdm==4.64.0
typing_extensions==4.1.1

-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torch==1.8.2+cu111
torchvision==0.9.2+cu111